🐛 [wire] Fix insert_bits truncation for fields wider than 32 bits
Problem:
- insert_msb and insert_lsb cast the value to unsigned (32-bit)
  before shifting, truncating fields wider than 32 bits
- 48-bit timestamp insert_bits/extract_bits round-trip produced wrong
  values in unoptimized (Debug) builds; Release masked the UB via
  optimization

Solution:
- Cast value to std::uint64_t before shifting in both insert_msb and
  insert_lsb
- Unit tests: 48-bit round-trip (MSB and LSB), large value (2^48-1),
  canary byte preservation

wide bits once lost—
sixty-four not thirty-two,
the shift now holds all

+++ src/include/xb/wire/bits.hpp
@@ -110,8 +110,10 @@ namespace xb::wire {
         unsigned chunk_width = hi - lo;
         bits_remaining -= chunk_width;
 
-        unsigned chunk = (static_cast<unsigned>(value) >> bits_remaining) &
-                         ((1u << chunk_width) - 1u);
+        unsigned chunk =
+            static_cast<unsigned>(
+                (static_cast<std::uint64_t>(value) >> bits_remaining)) &
+            ((1u << chunk_width) - 1u);
         unsigned mask = ((1u << chunk_width) - 1u) << (8 - hi);
         auto byte_val = static_cast<unsigned>(buf[i]);
         byte_val = (byte_val & ~mask) | (chunk << (8 - hi));
@@ -135,8 +137,10 @@ namespace xb::wire {
             (i + 1 == end_byte && end_bit % 8 != 0) ? (end_bit % 8) : 8;
         unsigned chunk_width = hi - lo;
 
-        unsigned chunk = (static_cast<unsigned>(value) >> bits_consumed) &
-                         ((1u << chunk_width) - 1u);
+        unsigned chunk =
+            static_cast<unsigned>(
+                (static_cast<std::uint64_t>(value) >> bits_consumed)) &
+            ((1u << chunk_width) - 1u);
         unsigned mask = ((1u << chunk_width) - 1u) << lo;
         auto byte_val = static_cast<unsigned>(buf[i]);
         byte_val = (byte_val & ~mask) | (chunk << lo);
